requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25

# HTTP & Scraping
requests==2.31.0
beautifulsoup4==4.12.3
aiohttp==3.9.1

# Discord
discord.py-self==2.0.0

# PDF Processing
pypdf2==3.0.1
pdfplumber==0.10.3

# AI APIs
anthropic==0.18.1
openai==1.12.0

# YouTube
google-api-python-client==2.116.0

# RSS
feedparser==6.0.11

# Utilities
python-dotenv==1.0.0
schedule==1.2.0
pydantic==2.5.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
